\name{as.graph.RHuginDomain}
\alias{as.graph.RHuginDomain}

\title{Coerce to Graph}

\description{
Coerce an RHugin domain to a graph.
}

\usage{
as.graph.RHuginDomain(x, ...)
}

\arguments{
  \item{x}{an RHugin domain.}
  \item{\dots}{additional arguments.}
}

\value{
  a representation of \code{x} as a graph (with class \sQuote{graphNEL}).
}

\author{Kjell Konis \email{kjell.konis@me.com}}

\keyword{programming}


