\name{generate.nodes}
\alias{generate.nodes}

\title{Generate Nodes}

\description{
Create nodes in an RHugin domain from the factor variables in a data frame.
}

\usage{
generate.nodes(domain, data)
}

\arguments{
  \item{domain}{an RHugin domain object.}
  \item{data}{a data frame containing one or more factor variables.}
}

\details{
Use the function \code{\link{generate.tables}} to set the initial conditional probability tables (CPT) for the nodes generated by this function.
}

\value{
a \code{NULL} value is invisibly returned.
}

\author{Kjell Konis \email{kjell.konis@epfl.ch}}

\examples{
library(MASS)
data(housing)

# Create an empty RHugin domain
hd <- hugin.domain()

# Generate nodes for the housing data
generate.nodes(hd, housing)
}

\keyword{programming}


