\name{print.RHuginDomain}
\alias{print.RHuginDomain}

\title{RHuginDomain Print Method}

\description{
  Print an RHuginDomain object.
}

\usage{
\method{print}{RHuginDomain}(x, ...)
}

\arguments{
  \item{x}{an RHuginDomain object.}
  \item{\dots}{additional arguments are ignored.}
}

\value{
  \code{x} is invisibly returned.
}

\author{Kjell Konis \email{kjell.konis@epfl.ch}}

\keyword{methods}


