\name{utilities}
\alias{RHugin.check.args}
\alias{.First.lib}

\title{Undocumented RHugin Utility Functions}

\description{
These functions are for RHugin's internal use only.
}

\usage{
RHugin.check.args(domain, required.nodes, conflicting.nodes, fun.name)
.First.lib(libname, pkgname)
}

\arguments{
  \item{domain}{domain}
  \item{required.nodes}{required.nodes}
  \item{conflicting.nodes}{conflicting.nodes}
  \item{libname}{libname}
  \item{pkgname}{pkgname}
  \item{fun.name}{fun.name}
}


\value{
Undocumented.
}

\author{Kjell Konis \email{kjell.konis@epfl.ch}}

\keyword{utilities}


