\name{as.graph}
\alias{as.graph}

\title{Coerce to a Graph}

\description{
Coerce an RHugin domain object to a graph.
}

\usage{
as.graph(x, ...)
}

\arguments{
  \item{x}{the object to be coerced.}
  \item{\dots}{additional arguments.}
}

\value{
  a representation of \code{x} as a graph (with class \sQuote{graphNEL}).
}

\author{Kjell Konis \email{kjell.konis@epfl.ch}}

\keyword{programming}


