\name{hugin.domain}
\alias{hugin.domain}

\title{Create an RHugin Domain Object}

\description{
Create an RHugin domain object.
}

\usage{
hugin.domain()
}

\value{
an RHugin domain object.
}

\references{
HUGIN API Reference Manual \url{http://www.hugin.com/developer/documentation/API_Manuals/}: \code{h_new_domain}.
}

\author{Kjell Konis \email{kjell.konis@epfl.ch}}

\examples{
# Create an empty RHugin domain
AppleTree <- hugin.domain()
}

\keyword{programming}


