\name{plot.RHuginDomain}
\alias{plot.RHuginDomain}

\title{Plot an RHugin Domain}

\description{
Plot method for RHugin domain objects.
}

\usage{
\method{plot}{RHuginDomain}(x, y, ...)
}

\arguments{
  \item{x}{an RHugin domain.}
  \item{y}{this argument is not used.}
  \item{\dots}{additional arguments are ignored.}
}

\details{
A graphical representation of the RHugin domain \sQuote{x} is generated by calling \code{\link{layoutRHugin}} and displayed on a graphics device.  Note that the call to \code{\link{layoutRHugin}} saves the generated layout in the RHugin domain.
}

\value{
the \sQuote{graph} object is invisibly returned.
}

\author{Kjell Konis \email{kjell.konis@epfl.ch}}

\examples{
library(Rgraphviz)

demo("buildChestClinicNetwork", echo = FALSE)
plot(ChestClinic)
write.rhd(ChestClinic, file = "ChestClinic.net", type = "net")

# The layout of the bayesian network in "ChestClinic.net" is the same as
# the layout in the R graphic device.
}

\keyword{hplot}


