\name{utilities}
\alias{RHugin.check.args}
\alias{RHugin.check.domain}
\alias{.First.lib}

\title{Undocumented RHugin Utility Functions}

\description{
These functions are for RHugin's internal use only.
}

\usage{
RHugin.check.args(domain, required.nodes, conflicting.nodes, fun.name)
RHugin.check.domain(domain, fun.name)
}

\arguments{
  \item{domain}{domain}
  \item{required.nodes}{required.nodes}
  \item{conflicting.nodes}{conflicting.nodes}
  \item{fun.name}{fun.name}
}


\value{
Undocumented.
}

\author{Kjell Konis \email{kjell.konis@me.com}}

\keyword{programming}


